I mean it works...?
The zombie resurrects after 10s. The original plan of teleport down 1.25 blocks during resurrection is not useful though.

Need to figure out some sort of *animation* for this to work as intended, or at least a placeholder.

TODO: Test stopZombieResurrects()
TODO: Configurable!
TODO: On disable, clear out all the zombies on resurrect.